better rendering for the cards and load their icons and started implementing game screen proportional to the size of the screen

+++ src/ui/Window.java
@@ -4,6 +4,8 @@ import java.awt.Rectangle;
 
 import javax.swing.JFrame;
 
+import core.gamestate.constants.GameRenderingConstants;
+
 public class Window extends JFrame {
     public static int screenWidth, screenHeight;
     public static float screenSizeUnit;
@@ -23,7 +25,11 @@ public class Window extends JFrame {
         Rectangle bounds = this.getBounds();
         Window.screenWidth = bounds.width;
         Window.screenHeight = bounds.height;
-        Window.screenSizeUnit = ((float) Window.screenWidth) / 100;
+        Window.screenSizeUnit = Window.screenWidth / 100;
+        // Window.screenSizeUnit = Window.screenWidth
+        // / Window.screenHeight > GameRenderingConstants.SCREEN_RATIO
+        // ? (float) (Window.screenHeight / GameRenderingConstants.HEIGHT_RATIO)
+        // : (float) (Window.screenWidth / GameRenderingConstants.WIDTH_RATIO);
     }
 
 }
